correct color scheme and add proper text sizes correct some model and made some ui better

+++ crm-get2uni/src/components/ui/calendar.tsx
@@ -35,7 +35,7 @@ function Calendar({
       captionLayout={captionLayout}
       formatters={{
         formatMonthDropdown: (date) =>
-          date.toLocaleString("default", { month: "short" }),
+          date.toLocaleString("default", { month: "long" }),
         ...formatters,
       }}
       classNames={{
@@ -123,6 +123,7 @@ function Calendar({
         ...classNames,
       }}
       components={{
+        
         Root: ({ className, rootRef, ...props }) => {
           return (
             <div
